feat(starter,docs): minimal landing with /docs redirect; add copyable Jinja include snippets across components; logo + env-configured links; adjust starter test
