+++ docs/components/toast.md
@@ -7,6 +7,13 @@
 - Accessibility: Each toast is `role="status"`; dismiss buttons include `aria-label`; icons are decorative `aria-hidden`.
 - Theming hooks: Override custom properties (`--_toast-*`) or variant colors to match brand guidelines.
 
+## Include in template
+
+```jinja
+{% include "greeble/toast.root.html" %}
+{% include "greeble/toast.item.html" %}
+```
+
 ## Server contract
 
 - Any endpoint can return a toast out-of-band by wrapping the item markup in a `<div id="greeble-toasts" hx-swap-oob="true">…</div>` container.
