+++ src/greeble_cli/templates/starter/starter_README.md
@@ -13,7 +13,7 @@ uv run greeble-starter
 uv run python -m greeble_starter
 ```
 
-Open http://127.0.0.1:8050/ to interact with the demo endpoints.
+Open http://127.0.0.1:8050/ to view the landing page.
 
 ## Structure
 
@@ -21,6 +21,19 @@ Open http://127.0.0.1:8050/ to interact with the demo endpoints.
   form, stepper, and infinite list endpoints
 - `templates/` – Base layout (`index.html`) and component templates copied from Greeble
 - `static/` – Core tokens (`greeble-core.css` copied via CLI) and a small site stylesheet
+
+## Configure links
+
+The landing page shows three buttons: Docs, GitHub, and Demo. You can change their targets with environment variables (read at runtime):
+
+```bash
+export GREEBLE_DOCS_URL="https://greeble-synai.ngrok.dev/docs"
+export GITHUB_REPO_URL="https://github.com/bakobiibizo/greeble"
+export GREEBLE_DEMO_URL="https://github.com/bakobiibizo/greeble/tree/release-candidate/examples"
+uv run greeble-starter --port 8050
+```
+
+This starter intentionally keeps endpoints minimal. Explore full component flows in the repository `examples/` if you need richer demos.
 
 ## Next steps
